requirements.txt
django==1.11.5
psycopg2==2.7.3.1
python-twitter==3.3
lxml==3.8.0
django-model-utils==3.0.0
djangorestframework==3.6.4
